🔥 Split logic in client into separate modules and create instructions for spl logic instead of manual account setups
